+++ src/main/kotlin/other/mvisetup/mviSetupTemplate.kt
@@ -1,15 +1,12 @@
 package other.mvisetup
 
 import com.android.tools.idea.wizard.template.Category
-import com.android.tools.idea.wizard.template.CheckBoxWidget
 import com.android.tools.idea.wizard.template.Constraint
 import com.android.tools.idea.wizard.template.FormFactor
 import com.android.tools.idea.wizard.template.ModuleTemplateData
-import com.android.tools.idea.wizard.template.PackageNameWidget
 import com.android.tools.idea.wizard.template.TemplateData
 import com.android.tools.idea.wizard.template.TextFieldWidget
 import com.android.tools.idea.wizard.template.WizardUiContext
-import com.android.tools.idea.wizard.template.booleanParameter
 import com.android.tools.idea.wizard.template.stringParameter
 import com.android.tools.idea.wizard.template.template
 
@@ -25,42 +22,41 @@ val mviSetupTemplate
             WizardUiContext.NewProject, WizardUiContext.NewModule
         )
 
-        val packageNameParam = defaultPackageNameParameter
         val className = stringParameter {
             name = "Class Name"
             default = "Test"
             help = "The name of the class to create and use in MVI and Fragment"
             constraints = listOf(Constraint.NONEMPTY)
         }
-        val generateLayout = booleanParameter {
-            name = "Generate a Layout File"
-            default = true
-            help = "If true, a layout file will be generated"
+
+        val packageName = stringParameter {
+            name = "Package name"
+            default = "com.emlid.reachview3.ui"
+            constraints = listOf(Constraint.NONEMPTY)
+            suggest = { packageName }
+        }
+
+        val containerName = stringParameter {
+            name = "Container name"
+            default = "SurveyContainer"
+            help = "The name of the container that will be used in router side effect"
+            constraints = listOf(Constraint.NONEMPTY)
         }
 
         widgets(
             TextFieldWidget(className),
-            PackageNameWidget(packageNameParam),
-            CheckBoxWidget(generateLayout)
+            TextFieldWidget(packageName),
+            TextFieldWidget(containerName)
         )
 
         recipe = { data: TemplateData ->
             mviSetup(
                 data as ModuleTemplateData,
-                packageNameParam.value,
+                packageName.value,
                 className.value,
-                generateLayout.value
+                containerName.value
             )
         }
     }
 
-val defaultPackageNameParameter
-    get() = stringParameter {
-        name = "Package name"
-        visible = { !isNewModule }
-        default = "com.emlid.reachview3.ui"
-        constraints = listOf(Constraint.PACKAGE)
-        suggest = { packageName }
-    }
-
-const val MIN_API = 16
+private const val MIN_API = 16