+++ src/main/kotlin/other/mvisetup/src/app_package/Fragment.kt
@@ -1,7 +1,6 @@
 package other.mvisetup.src.app_package
 
 import com.android.tools.idea.wizard.template.escapeKotlinIdentifier
-import other.mvisetup.camelToSnakeCase
 
 fun fragment(
     packageName: String,
@@ -9,73 +8,28 @@ fun fragment(
 ) = """package ${escapeKotlinIdentifier(packageName)}
 
 import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
-import com.emlid.reachview3.R
-import com.emlid.reachview3.databinding.Fragment${className}Binding
-import com.emlid.reachview3.mvibase.MviView
-import com.emlid.reachview3.navigation.BackButtonListener
-import com.emlid.reachview3.ui.custom.viewbinding.viewBinding
-import ${escapeKotlinIdentifier(packageName)}.mvi.${className}Intent
-import ${escapeKotlinIdentifier(packageName)}.mvi.${className}State
-import ${escapeKotlinIdentifier(packageName)}.mvi.${className}ViewModel
-import io.reactivex.Observable
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.rxkotlin.merge
-import io.reactivex.subjects.PublishSubject
+import android.view.ViewGroup
+import androidx.compose.ui.platform.ComposeView
+import com.emlid.reachview3.navigation.ComposeFragment
+import com.emlid.reachview3.navigation.ModalFragment
+import com.emlid.reachview3.ui.theme.ReachViewTheme
 import org.koin.androidx.scope.ScopeFragment
-import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class ${className}Fragment :
-    ScopeFragment(R.layout.fragment_${className.camelToSnakeCase()}),
-    MviView<${className}Intent, ${className}State>,
-    BackButtonListener {
+    ScopeFragment(),
+    ModalFragment,
+    ComposeFragment {
 
-    private val viewModel: ${className}ViewModel by viewModel()
-    private val viewBinding by viewBinding(Fragment${className}Binding::bind)
-
-    private val compositeDisposable = CompositeDisposable()
-    private val exitSubject: PublishSubject<${className}Intent.Exit> = PublishSubject.create()
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-    }
-
-    override fun onStart() {
-        super.onStart()
-        bind()
-    }
-
-    override fun bind() {
-        compositeDisposable.add(viewModel.states().subscribe(::render))
-        viewModel.subscribeToIntents(intents())
-    }
-
-    override fun render(state: ${className}State) {
-    }
-
-    override fun intents(): Observable<${className}Intent> =
-        listOf(
-            initialIntent(),
-            exitIntent()
-        ).merge()
-
-    private fun initialIntent(): Observable<${className}Intent> =
-        Observable.just(${className}Intent.Initial)
-
-    private fun exitIntent(): Observable<${className}Intent.Exit> =
-        exitSubject
-
-    override fun onStop() {
-        compositeDisposable.clear()
-        viewModel.unsubscribeFromIntents()
-        super.onStop()
-    }
-
-    override fun onBackPressed(): Boolean {
-        exitSubject.onNext(${className}Intent.Exit)
-        return true
-    }
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View =
+        ComposeView(inflater.context).apply {
+            setContent {
+                ReachViewTheme {
+                    ${className}Screen()
+                }
+            }
+        }
 
     companion object {