+++ src/main/kotlin/other/mvisetup/src/utils/StringUtils.kt
@@ -1,0 +1,9 @@
+package other.mvisetup.src.utils
+
+fun String.separateCamelCase(): String {
+    val camelRegex = "(?<=[a-zA-Z])[A-Z]".toRegex()
+
+    return camelRegex.replace(this) {
+        " ${it.value}"
+    }.toLowerCase().capitalize()
+}